Minor crc base implementation refactor to segregate that code to the NoOp DAL alone
